#144 - lib: split native function table from lib.rs
