+++ metrics/regression/current.summary
@@ -1,40 +1,40 @@
 Perf Metrics Summary: lib
 --------------------
-lib/compile        bytes:     1520     usecs:      48.82
-lib/core           bytes:     5808     usecs:     151.36
-lib/gcd            bytes:      624     usecs:     143.44
-lib/list           bytes:     5296     usecs:     119.10
-lib/namespace      bytes:      240     usecs:       4.02
-lib/number         bytes:     3600     usecs:      80.50
-lib/reader         bytes:     5280     usecs:     104.36
-lib/special-form   bytes:     2400     usecs:      60.68
-lib/stream         bytes:      480     usecs:      12.88
-lib/struct         bytes:      576     usecs:      13.54
-lib/symbol         bytes:     1200     usecs:      28.76
-lib/vector         bytes:     4456     usecs:     100.36
+lib/compile        bytes:     1520     usecs:      50.70
+lib/core           bytes:     5808     usecs:     164.10
+lib/gcd            bytes:      624     usecs:      97.80
+lib/list           bytes:     5296     usecs:      98.50
+lib/namespace      bytes:      240     usecs:       3.34
+lib/number         bytes:     3600     usecs:      64.00
+lib/reader         bytes:     5280     usecs:      85.34
+lib/special-form   bytes:     2400     usecs:      52.02
+lib/stream         bytes:      480     usecs:      11.22
+lib/struct         bytes:      576     usecs:      25.52
+lib/symbol         bytes:     1200     usecs:      23.46
+lib/vector         bytes:     4456     usecs:      86.30
 
 Perf Metrics Summary: frequent
 --------------------
-frequent/core      bytes:     9624     usecs:     636.00
-frequent/fixnum    bytes:     1680     usecs:      37.08
-frequent/float     bytes:     1200     usecs:      27.12
-frequent/list      bytes:     2160     usecs:      47.50
-frequent/vector    bytes:      240     usecs:       5.38
+frequent/core      bytes:     9624     usecs:     466.62
+frequent/fixnum    bytes:     1680     usecs:      29.32
+frequent/float     bytes:     1200     usecs:      21.42
+frequent/list      bytes:     2160     usecs:      37.86
+frequent/vector    bytes:      240     usecs:       4.46
 
 Perf Metrics Summary: prelude
 --------------------
-prelude/closure    bytes:    20904     usecs:   14568.96
-prelude/compile    bytes:     5512     usecs:    1570.06
-prelude/prelude    bytes:     4592     usecs:     269.30
-prelude/exception  bytes:     6896     usecs:    5200.42
-prelude/fixnum     bytes:     2000     usecs:     184.80
-prelude/format     bytes:     6184     usecs:    7046.12
-prelude/lambda     bytes:    97784     usecs:   66855.34
-prelude/list       bytes:    20864     usecs:    9010.64
-prelude/macro      bytes:    58640     usecs:   45224.56
-prelude/reader     bytes:    15216     usecs:  123656.44
-prelude/stream     bytes:      960     usecs:      71.18
-prelude/string     bytes:     2160     usecs:     595.90
-prelude/type       bytes:     8768     usecs:    6339.44
-prelude/vector     bytes:     9472     usecs:    6655.54
+prelude/closure    bytes:    20904     usecs:   10092.72
+prelude/compile    bytes:     5512     usecs:    1105.64
+prelude/prelude    bytes:     4592     usecs:     214.74
+prelude/exception  bytes:     6896     usecs:    3635.06
+prelude/fixnum     bytes:     2000     usecs:     147.44
+prelude/format     bytes:     6184     usecs:    4984.94
+prelude/lambda     bytes:    97784     usecs:   47562.08
+prelude/list       bytes:    20864     usecs:    6463.20
+prelude/macro      bytes:    58640     usecs:   31557.58
+prelude/reader     bytes:    15216     usecs:   83136.82
+prelude/stream     bytes:      960     usecs:      60.16
+prelude/string     bytes:     2160     usecs:     432.12
+prelude/type       bytes:     8768     usecs:    4466.96
+prelude/vector     bytes:     9472     usecs:    4756.42
