+++ src/libenv/features/nix/nix_.rs
@@ -7,7 +7,7 @@ use crate::{
         env::Env,
         exception::{self, Condition, Exception},
         frame::Frame,
-        lib::LibFnDef,
+        functions::LibFnDef,
         types::Tag,
     },
     features::Feature,
